Rename equinix package to github
